Seeded Key in PublicKeyBox
I noticed the GenerateKeyPair in PublicKeyBox class is inconsistent with the PublicKeyAuth class. Presently the private key is stored and the public key computed from it whereas in PublicKeyAuth the seed is used to compute a deterministic keypair by calling the underlying libsodium API. I have added a method GenerateSeededKeyPair to support this as I have used in other libsodium variants. Keen to understand if I have missed something here?